Don't create default value for `now` when inheriting in `NextIntlClientProvider`, adjust docs
