+++ packages/next-intl/src/react-server/NextIntlClientProviderServer.tsx
@@ -1,6 +1,7 @@
 import {ComponentProps} from 'react';
+import getConfigNow from '../server/react-server/getConfigNow.tsx';
 import getFormats from '../server/react-server/getFormats.tsx';
-import {getLocale, getNow, getTimeZone} from '../server.react-server.tsx';
+import {getLocale, getTimeZone} from '../server.react-server.tsx';
 import BaseNextIntlClientProvider from '../shared/NextIntlClientProvider.tsx';
 
 type Props = ComponentProps<typeof BaseNextIntlClientProvider>;
@@ -18,7 +19,8 @@ export default async function NextIntlClientProviderServer({
       // See https://github.com/amannn/next-intl/issues/631
       formats={formats === undefined ? await getFormats() : formats}
       locale={locale ?? (await getLocale())}
-      now={now ?? (await getNow())}
+      // Note that don't assign a default for `now` here
+      now={now ?? (await getConfigNow())}
       timeZone={timeZone ?? (await getTimeZone())}
       {...rest}
     />
