+++ packages/next-intl/src/server/react-server/getConfig.tsx
@@ -10,12 +10,6 @@ import {getRequestLocale} from './RequestLocale.tsx';
 import createRequestConfig from './createRequestConfig.tsx';
 import {GetRequestConfigParams} from './getRequestConfig.tsx';
 
-// Make sure `now` is consistent across the request in case none was configured
-function getDefaultNowImpl() {
-  return new Date();
-}
-const getDefaultNow = cache(getDefaultNowImpl);
-
 // This is automatically inherited by `NextIntlClientProvider` if
 // the component is rendered from a Server Component
 function getDefaultTimeZoneImpl() {
@@ -75,7 +69,6 @@ const getCache = cache(_createCache);
 async function getConfigImpl(localeOverride?: Locale): Promise<
   IntlConfig & {
     getMessageFallback: NonNullable<IntlConfig['getMessageFallback']>;
-    now: NonNullable<IntlConfig['now']>;
     onError: NonNullable<IntlConfig['onError']>;
     timeZone: NonNullable<IntlConfig['timeZone']>;
     _formatters: ReturnType<typeof _createIntlFormatters>;
@@ -88,13 +81,7 @@ async function getConfigImpl(localeOverride?: Locale): Promise<
   return {
     ...initializeConfig(runtimeConfig),
     _formatters: getFormatters(getCache()),
-    timeZone: runtimeConfig.timeZone || getDefaultTimeZone(),
-
-    // Only init when necessary to avoid triggering a `dynamicIO` error
-    // (i.e. when using `format.relativeTime` or `useNow`)
-    get now() {
-      return runtimeConfig.now ?? getDefaultNow();
-    }
+    timeZone: runtimeConfig.timeZone || getDefaultTimeZone()
   };
 }
 const getConfig = cache(getConfigImpl);
