+++ packages/next-intl/src/server/react-server/getConfigNow.tsx
@@ -1,0 +1,11 @@
+import {cache} from 'react';
+import type {Locale} from 'use-intl';
+import getConfig from './getConfig.tsx';
+
+async function getConfigNowImpl(locale?: Locale) {
+  const config = await getConfig(locale);
+  return config.now;
+}
+const getConfigNow = cache(getConfigNowImpl);
+
+export default getConfigNow;
