+++ packages/next-intl/src/server/react-server/getDefaultNow.tsx
@@ -1,0 +1,8 @@
+import {cache} from 'react';
+
+function getDefaultNowImpl() {
+  return new Date();
+}
+const getDefaultNow = cache(getDefaultNowImpl);
+
+export default getDefaultNow;
