+++ packages/next-intl/src/server/react-server/getNow.tsx
@@ -1,13 +1,7 @@
-import {cache} from 'react';
 import type {Locale} from 'use-intl';
-import getConfig from './getConfig.tsx';
-
-async function getNowCachedImpl(locale?: Locale) {
-  const config = await getConfig(locale);
-  return config.now;
-}
-const getNowCached = cache(getNowCachedImpl);
+import getConfigNow from './getConfigNow.tsx';
+import getDefaultNow from './getDefaultNow.tsx';
 
 export default async function getNow(opts?: {locale?: Locale}): Promise<Date> {
-  return getNowCached(opts?.locale);
+  return (await getConfigNow(opts?.locale)) ?? getDefaultNow();
 }
